refactor(objectstore): split presign provider modules
